Feat/add a box to show part details in search dailog (#20)
* add previous and next page

* feat(): add previous and next page

add previous page and next page

Remove timeouts which were used to mitigate the racing issue but are
obsolete now.

* feat(partselector.py):add previous and next page

This is to add previous and next page in BOM match data.You can click the button to book parts list.

* rebase

* feat(partselector.py):add previous and next page  button

This is to add previous and next page button in BOM match.You can click the button to book parts list.

* feat(partselector.py):add previous and next page  button

This is to add previous and next page button in BOM match.You can click the button to book parts list.

* feat: add previous and next page  button

This is to add previous and next page button in BOM match. You can click the button to  check parts list.

* feat: add previous and next page  buttons

The purpose is to incorporate previous and next page buttons into 'NextPCB Search Online'. You can click these buttons to browse the parts list.

* feat: add previous and next page  buttons

The purpose is to incorporate previous and next page buttons into 'NextPCB Search Online'. You can click these buttons to browse the parts list.

* feat: add previous and next page buttons

* feat: add previous and next page buttons

* Feat/add a box to show part details in search dailog